LayoutTests/http/tests/security/contentSecurityPolicy/report-uri-from-javascript-expected.txt

CONSOLE MESSAGE: Refused to apply inline style because it violates the following Content Security Policy directive: "style-src 'none'".

CONSOLE MESSAGE: Refused to apply inline style because it violates the following Content Security Policy directive: "style-src 'none'".

CSP report received:
CONTENT_TYPE: application/json
HTTP_REFERER: http://127.0.0.1:8000/security/contentSecurityPolicy/report-uri-from-javascript.html
REQUEST_METHOD: POST
=== POST DATA ===
{"csp-report":{"document-uri":"http://127.0.0.1:8000/security/contentSecurityPolicy/report-uri-from-javascript.html","referrer":"","violated-directive":"style-src 'none'","original-policy":"style-src 'none'; report-uri resources/save-report.php","blocked-uri":"","source-file":"http://127.0.0.1:8000/security/contentSecurityPolicy/resources/inject-style.js","line-number":5}}
